Convert the Trees JavaScript to TypeScript
