+++ betty/extension/trees/assets/betty.extension.npm._Npm/src/package.json
@@ -1,5 +1,5 @@
 {
-  "main": "trees.js",
+  "main": "trees.ts",
   "engines": {
     "node": ">= 16"
   },
