+++ package.json
@@ -5,6 +5,8 @@
   "devDependencies": {
     "@playwright/test": "^1.43.1",
     "@stylistic/eslint-plugin": "^1.7.2",
+    "@types/cytoscape": "^3.21.0",
+    "@types/cytoscape-dagre": "^2.3.3",
     "@types/node": "*",
     "betty.extension.CottonCandy":  "file:./betty/extension/cotton_candy/assets/betty.extension.npm._Npm/src",
     "betty.extension.HttpApiDoc":  "file:./betty/extension/http_api_doc/assets/betty.extension.npm._Npm/src",
